+++ spec/support/custom_collector.rb
@@ -1,4 +1,4 @@
-# Copyright (c) 2017-present, BigCommerce Pty. Ltd. All rights reserved
+# Copyright (c) 2020-present, BigCommerce Pty. Ltd. All rights reserved
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
 # documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
@@ -13,20 +13,26 @@
 # COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
 # OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #
-require_relative 'rpc/ThingService_services_pb'
-require_relative 'rpc/Error_pb'
-
-class ThingController < ::Gruf::Controllers::Base
-  bind ::Rpc::ThingService::Service
-
-  def get_thing
-    collector.bar_it_up(total: rand(1_000))
-    Rpc::GetThingResponse.new(thing: Rpc::Thing.new(id: request.message.id, name: 'Johnny'))
+class CustomCollector < ::Bigcommerce::Prometheus::Collectors::Base
+  ##
+  # A demo metric on-demand
+  #
+  def bar_it_up(total:)
+    metric = {}
+    metric[:type] = 'custom'
+    metric[:custom_labels] = {}
+    metric[:on_demand_total] = total.to_i
+    @logger.debug("[gruf-prometheus] Pushing custom metric on_demand_total to type collector #{Bigcommerce::Prometheus.server_host}:#{Bigcommerce::Prometheus.server_port}: #{metric.inspect}")
+    @client.send_json(metric)
+  rescue StandardError => e
+    @logger.error("[gruf-prometheus] Prometheus failed to send on_demand_total stats: #{e.message}")
   end
 
-  private
-
-  def collector
-    @collector ||= ::CustomCollector.new
+  ##
+  # Metrics to be collected on a polling basis
+  #
+  def collect(metrics)
+    metrics[:polled_total] = rand(100)
+    metrics
   end
 end